+++ .gitignore
@@ -1,2 +1,2 @@
-appEnv/
+vEnv/
 __pycache__/
